requirements.txt
absl-py==1.0.0
altair==4.2.2
ansi2html==1.8.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.5
Bottleneck==1.3.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==37.0.1
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
Faker==8.8.1
fastjsonschema==2.15.1
Flask==2.2.2
freezegun==1.1.0
gitdb==4.0.10
GitPython==3.1.30
google-ai-generativelanguage==0.3.3
google-api-core==2.14.0
google-auth==2.23.4
google-generativeai==0.2.2
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.9.1
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-dash==0.4.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.2
mdurl==0.1.2
mistune==3.0.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nbclient==0.5.13
nbconvert==7.7.2
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numexpr==2.8.4
numpy==1.22.3
ortools==9.3.10459
packaging==21.3
pandas==1.3.1
pandocfilters==1.5.0
paramiko==2.8.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
pipreqs==0.4.11
plotly==5.13.0
prometheus-client==0.13.1
prompt-toolkit==3.0.20
proto-plus==1.22.3
protobuf==3.20.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
PyNaCl==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.0
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pyTigerDriver==1.0.15
pyTigerGraph==1.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty==2.0.2
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests==2.28.1
retrying==1.3.4
rich==13.3.1
rsa==4.9
semver==2.13.0
Send2Trash==1.8.0
setuptools==65.5.0
sip==4.19.13
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
stack-data==0.2.0
streamlit==1.28.2
tenacity==8.2.0
terminado==0.13.1
testpath==0.6.0
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
tornado==6.1
tqdm==4.66.1
traitlets==5.1.1
typing_extensions==4.8.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==3.5.2
wincertstore==0.2
yarg==0.1.9
zipp==3.12.1
